generate_riddle,generate_hint関数の切り出し
